improve JSON docs
